requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
InquirerPy>=0.3.4
rich>=13.7.0
openpyxl>=3.1.0
pyarrow>=15.0.0
